refact: refatorações shellcheck

+++ src/runtimes/nodejs/uninstall.sh
@@ -6,17 +6,17 @@ readonly version="${2:-"all"}"
 
 uninstall_all() {
     # Apagar todos os arquivos de instalação do runtime especificado
-    rm -rf "${CLIVERMAN_INSTALLS_PATH}/${name}"
+    rm -rf "${CLIVERMAN_INSTALLS_PATH:?}/${name}"
     
     # Apagar o arquivo current version do runtime especificado
-    rm -f "${CLIVERMAN_INSTALLS_PATH}/current_versions/${name}.txt"
+    rm -f "${CLIVERMAN_INSTALLS_PATH:?}/current_versions/${name}.txt"
     
     # Apagar todos os shims relacionados ao runtime especificado
-    rm -f "${CLIVERMAN_SHIMS_PATH}/npm"
-    rm -f "${CLIVERMAN_SHIMS_PATH}/npx"
-    rm -f "${CLIVERMAN_SHIMS_PATH}/node"
-    rm -f "${CLIVERMAN_SHIMS_PATH}/corepack"
-    
+    rm -f "${CLIVERMAN_SHIMS_PATH:?}/npm"
+    rm -f "${CLIVERMAN_SHIMS_PATH:?}/npx"
+    rm -f "${CLIVERMAN_SHIMS_PATH:?}/node"
+    rm -f "${CLIVERMAN_SHIMS_PATH:?}/corepack"
+
     echo -e "\033[91m ${name}:all\033[0m"
 }
 
@@ -30,14 +30,14 @@ uninstall_version() {
     
     if [[ "$current_version" == "${version}" ]]; then
         # Apagar todos os shims relacionados ao runtime especificado
-        rm -f "${CLIVERMAN_SHIMS_PATH}/npm"
-        rm -f "${CLIVERMAN_SHIMS_PATH}/npx"
-        rm -f "${CLIVERMAN_SHIMS_PATH}/node"
-        rm -f "${CLIVERMAN_SHIMS_PATH}/corepack"
-        rm -f "${CLIVERMAN_INSTALLS_PATH}/current_versions/${name}"
+        rm -f "${CLIVERMAN_SHIMS_PATH:?}/npm"
+        rm -f "${CLIVERMAN_SHIMS_PATH:?}/npx"
+        rm -f "${CLIVERMAN_SHIMS_PATH:?}/node"
+        rm -f "${CLIVERMAN_SHIMS_PATH:?}/corepack"
+        rm -f "${CLIVERMAN_INSTALLS_PATH:?}/current_versions/${name}"
     fi
     
-    rm -rf "${CLIVERMAN_INSTALLS_PATH}/${name}/${version}"
+    rm -rf "${CLIVERMAN_INSTALLS_PATH:?}/${name}/${version}"
     
     echo -e "\033[91m ${name}:${version}\033[0m"
 }
